refactor: refactor KitProvider
