+++ packages/components/src/hocs/Provider/hooks/useStatusBarTheme/index.native.ts
@@ -1,0 +1,29 @@
+import { useLayoutEffect } from 'react';
+
+import { StatusBar } from 'react-native';
+
+import type { IUseStatusBarTheme } from './type';
+
+const setLightContent = (isAnimated = true) => {
+  StatusBar.pushStackEntry({
+    animated: isAnimated,
+    barStyle: 'light-content',
+  });
+};
+
+const setDarkContent = (isAnimated = true) => {
+  StatusBar.pushStackEntry({
+    animated: isAnimated,
+    barStyle: 'dark-content',
+  });
+};
+
+export const useStatusBarTheme: IUseStatusBarTheme = (themeVariant) => {
+  useLayoutEffect(() => {
+    if (themeVariant === 'light') {
+      setDarkContent();
+    } else if (themeVariant === 'dark') {
+      setLightContent();
+    }
+  }, [themeVariant]);
+};
