+++ packages/components/src/hocs/Provider/hooks/useStatusBarTheme/index.ts
@@ -1,0 +1,3 @@
+import type { IUseStatusBarTheme } from './type';
+
+export const useStatusBarTheme: IUseStatusBarTheme = () => {};
