+++ packages/components/src/hocs/Provider/hooks/useStatusBarTheme/type.ts
@@ -1,0 +1,1 @@
+export type IUseStatusBarTheme = (themeVariant: 'light' | 'dark') => void;
