+++ packages/components/src/layouts/Navigation/utils/StatusBarUtils.native.ts
@@ -20,16 +20,3 @@ export const setHidden = (
   StatusBar.setHidden(hidden, animation);
 };
 
-export const setLightContent = (isAnimated = true) => {
-  StatusBar.pushStackEntry({
-    animated: isAnimated,
-    barStyle: 'light-content',
-  });
-};
-
-export const setDarkContent = (isAnimated = true) => {
-  StatusBar.pushStackEntry({
-    animated: isAnimated,
-    barStyle: 'dark-content',
-  });
-};
